Move init tests to own file
